[RBDI-9] Added VehicleManager singleton to work with ConfigDataManager and BackendManager singletons.
Cherry-picked the Vehicle data model class from RBDI-6. Added a singleton VehicleManager class to wrap the BackendServer singleton and manage the vehicle data and subscription stuff being sent to/from the server. It now watches the Config singleton for vehicle id changes and the backend server for connection status to send subscriptions for default signals. It will contain and control the Vehicle instance that the view controllers can do KVO on.