[RBDI-9] Added lots and lots of packet handling code.
Added the packet parser class from the RVI core iOS SDK. Added a bunch of packet classes to encapsulate the handling of the network data and json stuff. Updated the parser for the new objects, and having the VehicleManager class handling all of that. Of course, can't test it; backend doesn't exist.