[RBDI-11] First steps in fleshing out the vehicle class.
Made all the properties type Signal. Added a mapping between the property names and the signals as known by the JLR vehicle, encapsulated in the vehicle object. Added a method that uses KVC to update the signal properties of the vehicle object given the JLR vehicle signal string, caching it if the signal data isn't downloaded yet. Need to fetch signal data for the vehicle when first initialized. Need KVO in UI for when signal data changes and signal values change.