[RBDI-11] More work on signal descriptor request.
Added a packet type for errors. Commented out the unneeded json processing from the packet parser code. Probably the whole idea of data binding into packets is a bit extra work, now that using websockets, but still helps with encapsulation in the long run and it's already done. Added some code to SignalManager but not sure where I want to go yet. Need to flesh out the API.