[RBDI-11] Cleaning up my app's logical state machine code.
Added a status packet type. Finished removing the json parsing/delegate code from the packet parser class and moved packetToString to that class. Updated BSM a bit with the changes to PacketParser. Updated the flow of VM w the new status packet. Updated Vehicle to not ask for descriptors when vid is null, like when vid is being changed.